+++ src/main/java/chess/controller/ChessController.java
@@ -5,8 +5,8 @@ import chess.dto.ChessPieceDto;
 import chess.dto.CurrentTurnDto;
 import chess.dto.ErrorResponseDto;
 import chess.dto.MoveRequestDto;
-import chess.result.EndResult;
-import chess.result.MoveResult;
+import chess.domain.result.EndResult;
+import chess.domain.result.MoveResult;
 import chess.service.ChessService;
 import chess.service.RoomService;
 import java.util.List;
