+++ src/main/java/chess/dao/ChessPieceDao.java
@@ -3,7 +3,6 @@ package chess.dao;
 import chess.domain.chesspiece.ChessPiece;
 import chess.domain.position.Position;
 import chess.dto.ChessPieceDto;
-import chess.dto.ChessPieceMapper;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
