+++ src/main/java/chess/domain/chessboard/ChessBoard.java
@@ -5,7 +5,7 @@ import chess.domain.Score;
 import chess.domain.chesspiece.ChessPiece;
 import chess.domain.chesspiece.Color;
 import chess.domain.position.Position;
-import chess.result.MoveResult;
+import chess.domain.result.MoveResult;
 import java.util.Map;
 import java.util.Objects;
