+++ src/main/java/chess/service/ChessService.java
@@ -11,13 +11,13 @@ import chess.domain.chesspiece.ChessPiece;
 import chess.domain.chesspiece.Color;
 import chess.domain.position.Position;
 import chess.dto.ChessPieceDto;
-import chess.dto.ChessPieceMapper;
+import chess.dao.ChessPieceMapper;
 import chess.dto.CurrentTurnDto;
 import chess.dto.MoveRequestDto;
 import chess.dto.RoomStatusDto;
-import chess.result.EndResult;
-import chess.result.MoveResult;
-import chess.result.StartResult;
+import chess.domain.result.EndResult;
+import chess.domain.result.MoveResult;
+import chess.domain.result.StartResult;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
