--- src/main/resources/application.properties
@@ -1,6 +1,0 @@
-spring.h2.console.enabled=true
-
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.datasource.url=jdbc:mysql://localhost:3306/chess
-spring.datasource.username=root
-spring.datasource.password=root
