+++ src/test/java/chess/dao/ChessPieceMapperTest.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
